feat : 수동 로또 개수 클래스 네이밍 수정, 부생성자 추가
